Feat: software config adapted

+++ src/semaforo.c
@@ -1,9 +1,8 @@
 #include "semaforo.h"
-#include <stdint.h>
 
 Semaforo Semaforo_new(uint8_t id, EstadoSemaforo estado, Direccion direccion,
-                      uint8_t duracion_verde, uint8_t duracion_amarillo,
-                      uint8_t duracion_rojo, uint8_t timer,
+                      uint16_t duracion_verde, uint16_t duracion_amarillo,
+                      uint16_t duracion_rojo, uint16_t timer,
                       Coordenada posicion) {
     return (Semaforo){
         .id = id,
